Transceiver52M: hack for transmitter disable with power off command
This patch will simply disable the tx burst submission at the output of
the filler table when the POWEROFF command is issued.

A more complicated patch would involve thread control or resource deallocation.
Basically, we can do something cheap and simple like this patch, or something
that deals with the existing shutdown/resource management issues. The latter
requires much more effort and thus we pospone it.